refactor: onChange 대신 onBlur 하여 렌더링 측면 개선
